[#2451] Refactor: Move utils columns related to utilities/column.js
